DBZ-7717 Move LogPositionValidator outside JdbcConnection
